Add hack for bundled JDK-8
This is a hack to circumvent issues where the version of a system JDK
and the bundled JDK does not match. Since the paths to libjvm depend
on the Java version, we have to do some guessing which Java version
will be used. This guessing is for now based on the system JDK, so
that we might end up looking for a path (based on the system JDK)
that does not exist in the bundled JDK.

This commit adds a HACK that forces the detection of the bundled JDK
to look for JDK-8 folder structures instead of the ones implied by the
system JDK.